fixed bugs

+++ eagle/src/main/scala/com/eagle/dao/entity/AdToEmbed.scala
@@ -3,7 +3,7 @@ package com.eagle.dao.entity
 /**
  * Created by Achia.Rifman on 10/01/2015.
  */
-case class AdToEmbed(val adPath :String, val startTime : Int, val endTime : Int, val cornerType : String, val cornerWidth : Int,
+case class AdToEmbed(val adPath :String, val startTime : Long, val endTime : Long, val cornerType : String, val cornerWidth : Int,
                       val cornerHeight : Int) {
 
 }
